requirements.txt
flask==3.1.1
flask-cors==5.0.1
google-generativeai==0.8.5
python-dotenv==1.1.0
requests==2.31.0
beautifulsoup4==4.12.2
gunicorn==22.0.0